refactor(GPS): split GPS services, isolate RTK, and add Qt integrations
Composition:
- GPSManager owns QGCPositionManager, which is no longer a separate
  singleton; correction state and the NTRIP GGA sources move to
  GPSCorrectionStatus and GPSGgaSources.
- NTRIPStreamSession owns one connection's transport, correction
  registration, GGA and stats; NTRIPManager keeps state and policy.
- NTRIP reconnects wait while the host has no network, through an injected
  QNetworkInformation monitor; loopback casters are never gated.

RTK:
- src/GPS/RTK builds as QGCGPSRTK. GPSRtk publishes a plain Status that an
  app-side GPSRTKFactGroup mirrors into Facts; QML reads
  gpsManager.gpsRtkFacts. Serial ports go through a GPSSerialPorts
  interface with a SerialPortManager adapter.
- RTKConnectionPolicy returns a structured session outcome; GPSRtk builds
  all user messages. TCP and UDP bridges share one BRIDGE_BAUDRATE.

Drivers and corrections:
- A GPSConfigurationSequence step engine runs SBF and Ashtech
  configuration; commands on the wire are unchanged.
- Rename the protocol classes and files to UBXProtocol, SBFProtocol and
  the like, with GPSDecoded* report types and QGCGPSProtocols targets.
- One move-only correction source registration replaces the separate
  token, with a copyable Weak for queued callbacks.

Lifetime and diagnostics:
- Observers of a GPS object's change notifications must use deleteLater();
  deleting the owner from a notification logs a warning.
- Log enum names and the applied configurations, with the NTRIP password
  and fixed-base coordinates redacted.
- Name the remaining freshness thresholds; drop an unused thread
  interruption request.

Build and tests:
- QGCGPSCore links QGCGPSReceiver privately; the MAVLink packet library
  merges into QGCGPSRTCMMavlink; RTK gets the components gate.
- Protocol tests share an explicit GPSTestClock instead of a compile
  definition that changed shared inline headers per source file.
- Remove the remaining license headers from GPS sources.

+++ src/GPS/NTRIP/NTRIPHttpSession.cc
@@ -46,7 +46,7 @@ void NTRIPHttpSession::_attach(QTcpSocket* socket, bool allowSelfSignedCerts)
             return;
         }
         const QString message = _socket->errorString();
-        qCWarning(NTRIPHttpSessionLog) << "Socket error code:" << static_cast<int>(code) << "msg:" << message;
+        qCWarning(NTRIPHttpSessionLog) << "Socket error code:" << code << "msg:" << message;
         _fail(NTRIPError::SocketError, message);
     });
     connect(socket, &QTcpSocket::readyRead, this, &NTRIPHttpSession::_read);
